requirements.txt
web3==5.21.0
python-dotenv==0.19.0